+++ packages/components/icon/src/Icon.tsx
@@ -15,14 +15,21 @@ export interface IconProps extends IconVariantsProps, React.SVGProps<SVGElement>
   label?: string
 }
 
-export function Icon({ label, className, size = 'md', children, ...others }: IconProps) {
+export function Icon({
+  label,
+  className,
+  size = 'current',
+  intent = 'current',
+  children,
+  ...others
+}: IconProps) {
   const child = React.Children.only(children)
 
   return (
     <>
       {React.cloneElement(child as React.ReactElement, {
+        className: iconVariants({ className, size, intent }),
         'data-spark-component': 'icon',
-        className: iconVariants({ className, size }),
         'aria-hidden': 'true',
         focusable: 'false',
         ...others,
@@ -32,3 +39,5 @@ export function Icon({ label, className, size = 'md', children, ...others }: Ico
     </>
   )
 }
+
+Icon.displayName = 'Icon'