+++ packages/utils/cli/src/generate/templates/util/[package.json].js
@@ -10,6 +10,16 @@ export default ({ name, description }) => `{
   "types": "./dist/index.d.ts",
   "scripts": {
     "build": "vite build"
-  }
+  },
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:adevinta/spark.git",
+    "directory": "packages/utils/${name}"
+  },
+  "bugs": {
+    "url": "https://github.com/adevinta/spark/issues"
+  },
+  "homepage": "https://sparkui.vercel.app",
+  "license": "MIT"
 }
 `