+++ packages/utils/internal-utils/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.3](https://github.com/adevinta/spark/compare/@spark-ui/internal-utils@1.5.2...@spark-ui/internal-utils@1.5.3) (2023-04-28)
+
+**Note:** Version bump only for package @spark-ui/internal-utils
+
+## [1.5.2](https://github.com/adevinta/spark/compare/@spark-ui/internal-utils@1.5.1...@spark-ui/internal-utils@1.5.2) (2023-04-24)
+
+### Bug Fixes
+
+- **internal-utils:** props table ([3552bc8](https://github.com/adevinta/spark/commit/3552bc8ec8b7012a2cbdb74317f9fb531e43f585))
+
+## [1.5.1](https://github.com/adevinta/spark/compare/@spark-ui/internal-utils@1.5.0...@spark-ui/internal-utils@1.5.1) (2023-04-19)
+
+**Note:** Version bump only for package @spark-ui/internal-utils
+
+# [1.5.0](https://github.com/adevinta/spark/compare/@spark-ui/internal-utils@1.4.0...@spark-ui/internal-utils@1.5.0) (2023-04-19)
+
+### Features
+
+- **internal-utils:** add current values and make variants partial ([5b1eb81](https://github.com/adevinta/spark/commit/5b1eb817fcbf1f05143e919fb4d37864828c257c))
+
 # [1.4.0](https://github.com/adevinta/spark/compare/@spark-ui/internal-utils@1.3.0...@spark-ui/internal-utils@1.4.0) (2023-04-06)
 
 ### Features